lareserve_backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
asgiref==3.11.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.4.22
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
cryptography==48.0.0
    # via lareserve (pyproject.toml)
dj-rest-auth==7.2.0
    # via lareserve (pyproject.toml)
django==6.0.5
    # via
    #   lareserve (pyproject.toml)
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
    #   drf-spectacular
django-allauth==65.16.1
    # via lareserve (pyproject.toml)
django-cors-headers==4.9.0
    # via lareserve (pyproject.toml)
django-extensions==4.1
    # via lareserve (pyproject.toml)
djangorestframework==3.17.1
    # via
    #   lareserve (pyproject.toml)
    #   dj-rest-auth
    #   drf-spectacular
drf-spectacular==0.29.0
    # via lareserve (pyproject.toml)
idna==3.13
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
loguru==0.7.3
    # via lareserve (pyproject.toml)
psycopg2-binary==2.9.12
    # via lareserve (pyproject.toml)
pycparser==3.0
    # via cffi
pyjwt==2.12.1
    # via lareserve (pyproject.toml)
python-dotenv==1.2.2
    # via lareserve (pyproject.toml)
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via lareserve (pyproject.toml)
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.5
    # via django
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.6.3
    # via requests
